refactor: [#1493] remove duplicate code for Peer buidler
